Loosen up WKService / WKServiceRequest
- create more specific WKMediaWikiServiceRequest to handle MediaWiki requests
- Allows WKService / WKServiceRequest to be used for alternative backend services
