+++ Sources/WKData/Service/WKService.swift
@@ -1,6 +1,6 @@
 import Foundation
 
 public protocol WKService {
-    func perform(request: WKServiceRequest, tokenType: WKServiceRequest.TokenType?, completion: @escaping (Result<[String: Any]?, Error>) -> Void)
-    func performDecodableGET<T: Decodable>(request: WKServiceRequest, completion: @escaping (Result<T, Error>) -> Void)
+    func perform<R: WKServiceRequest>(request: R, completion: @escaping (Result<[String: Any]?, Error>) -> Void)
+    func performDecodableGET<R: WKServiceRequest, T: Decodable>(request: R, completion: @escaping (Result<T, Error>) -> Void)
 }
