+++ Sources/WKData/Service/WKServiceRequest.swift
@@ -1,29 +1,15 @@
 import Foundation
 
-public struct WKServiceRequest {
-    public enum TokenType {
-        case csrf
-        case watch
-        case rollback
-    }
+public enum WKServiceRequestMethod: String {
+    case GET
+    case POST
+    case PUT
+    case DELETE
+    case HEAD
+}
 
-	public enum Method: String {
-		case GET
-		case POST
-		case PUT
-		case DELETE
-		case HEAD
-	}
-
-	public let url: URL?
-	public let method: Method
-    public let tokenType: TokenType?
-	public let parameters: [String: Any]?
-
-    internal init(url: URL? = nil, method: WKServiceRequest.Method, tokenType: WKServiceRequest.TokenType? = nil, parameters: [String : Any]? = nil) {
-        self.url = url
-        self.method = method
-        self.tokenType = tokenType
-        self.parameters = parameters
-    }
+public protocol WKServiceRequest {
+    var url: URL? { get }
+    var method: WKServiceRequestMethod { get }
+    var parameters: [String: Any]? { get }
 }
